requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-allauth==0.44.0
django-filter==2.4.0
django-ipware==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
html5lib==1.0.1
idna==2.8
install==1.3.4
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==0.6.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
simplejson==3.17.2
six==1.14.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
